Add files via upload
Sentiment analysis is a popular task in natural language processing. 
-> The goal of sentiment analysis is to classify the text based on the mood or mentality expressed in the text, which can be positive negative, or neutral.
Understanding Public Opinion: 
Analyzing sentiments expressed on platforms like Reddit helps in understanding the public's opinion and emotional responses toward a specific topic, such as cancer. This information can be useful for researchers, healthcare professionals, and policymakers to gauge how people perceive and discuss the subject.

Identifying Trends and Issues
By analyzing sentiment, one can identify emerging trends, concerns, or issues within the cancer community on Reddit. This information can be crucial for healthcare organizations, advocacy groups, and researchers to address specific challenges, gaps in information, or areas where more support is needed.

Overall, sentiment analysis on topics like cancer from Reddit and other social platforms can contribute to a deeper understanding of the human experience, inform decision-making, and facilitate the development of more targeted and empathetic approaches in healthcare, research, and support services

Stop words:
Stop words are common words in any language that occur with a high frequency but carry much less substantive information about the meaning of a phrase.
Examples of some common stop words include:
a, the, and , or , of , on , this , we , were, is, not …

In natural language processing (NLP), stopwords are frequently filtered out to enhance text analysis and computational efficiency. Eliminating stopwords can improve the accuracy and relevance of NLP tasks by drawing attention to the more important words, or content words.