Add files via upload
TFIDF

In information retrieval, the “term frequency – inverse document frequency” (also called TFIDF), is a well know method to evaluate how important is a word in a document.  TFIDF comes up a lot in research work because it’s both a corpus exploration method and a pre-processing step for many other text-mining measures and models. 

To get TFIDF we need 4 data structures:
- A list of all keywords in the corpus (K). This list is initially empty and grows are tweets are read and processed.
- A one-dimensional list of all words in a certain document (W). This structure is used for storing the words of a document as it is being processed. 
- The two-dimensional sparse array (TF) for storing the TF and/or TFIDF: the number of rows is the number of documents, and the columns represent the set of all words used in the corpus. 
- A one-dimentional list of document frequencies (DF), which is a parallel list of K and has the same size, it has the count of how many document contain that word. 

Dataset: un-general-debates-blueprint.csv