Add files via upload
Sentiment Analysis

Sentiment Analysis is a use case of Natural Language Processing (NLP) and comes under the category of text classification. To put it simply, Sentiment Analysis involves classifying a text into various sentiments, such as positive or negative, Happy, Sad or Neutral, etc. Thus, the ultimate goal of sentiment analysis is to decipher the underlying mood, emotion, or sentiment of a text. This is also known as Opinion Mining.

ARTS AND CRAFTS SUPPLIES DATASET OVERVIEW

The dataset consists of 494,485 reviews sourced from Amazon, focusing on Arts and Crafts supplies. The dataset contains several columns:

Columns:

1. Overall: This column represents the grade given by the reviewer, ranging from 1 (lowest) to 5 (highest), indicating the overall satisfaction level with the product.

2. Verified: This column indicates whether the purchase of the product was verified by Amazon, serving as a measure of the review's authenticity.

3. Reviewer ID: Each reviewer is assigned a unique identifier, allowing for tracking and analysis of individual reviewers.

4. ASIN: This column contains a unique product number assigned to each product, facilitating product-specific analysis.

5. Text: This column contains the full text of the review, providing detailed feedback from the reviewer.

6. Summary: This column contains a headline or summary of the review, offering a concise overview of the reviewer's sentiments.