Add files via upload
TEXT SUMMARIZATION CLASSWORK-5

NLP text summarization is the process of breaking down lengthy text into digestible paragraphs or sentences. This method extracts vital information while also preserving the meaning of the text. This reduces the time required for grasping lengthy pieces such as articles without losing vital information.

Text summarization is the process of creating a concise, coherent, and fluent summary of a longer text document, which involves underlining the document’s key points.

There are two approaches to text summarization.

Extractive approaches
Abstractive approaches

Text summarization uses AI and NLP to take large amounts of data and extract the most relevant details from the text.
The most important points within the document are summarized, while the meaning of the data is maintained.

Two Approaches to NLP Text Summarization
Extractive Text Summarization Approaches
With these approaches, we use an extractive technique to summarize our material using algorithms.

For example, if we want to summarize our text using the frequency technique, we store all of the key terms as well as the frequency of all of those words in a dictionary.

We record the sentences containing those terms in our final summary based on high-frequency words.

This ensures that the words in our summary come from the text.

The extraction is carried out in accordance with the specified measure, with no changes made to the text. See the graphic below for a description of the extractive summarization process.

To sum it up, this method detects essential pieces of text, cuts them out, and then stitches them back together to generate a condensed or concentrated form.

Abstractive Text Summarization Approaches
Abstractive summarizers do not produce the summary by selecting sentences from the original text passage; instead, they create a paraphrase of the key points of a given text, using a set of words that differs from the original.

To recap, this is quite similar to what we do as humans. In our minds, we form a phonological representation of a document.

We then select terms from our general vocabulary (words we frequently use) that fit in the semantics to construct a brief summary that conveys all of the document’s meaning.

In contrast to the extraction approach, where we only use the terms that are present – the sentence should not end here, it’s missing the main point of the sentence! (In contrast to this method does this and that) . It is highly likely that the abstractive summary phrases do not exist in the original text.

Below are the processes for extractive and abstractive summarization.

Six Benefits of NLP Text Summarization
1. Scalable and Quick
2. Leverages Existing Tools
3. Understand Your Customers Better
4. Summarize the Text in Different Formats
5. Ensure all Critical Information is Covered
6. Run Sentiment Analysis